Made different projects for each existing class
