+++ FirstLibrary/FirstLibrary/Person.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using Job;
 
 namespace FirstLibrary
 {
@@ -6,27 +7,13 @@ namespace FirstLibrary
     {
         public string name;
         public int age;
-        public Job job;
+        public Job.Job job;
 
-        public Person(string name, int age, Job job) // Consructor
+        public Person(string name, int age, Job.Job job) // Consructor
         {
             this.name = name;
             this.age = age;
             this.job = job;
         }
     }
-
-    public class Job
-    {
-        public string jobName;
-        public double jobSalary;
-
-        public Job(string jobName, double jobSalary) // Consructor
-        {
-            this.jobName = jobName;
-            this.jobSalary = jobSalary;
-        }
-    }
-
-
 }
