+++ FirstLibrary/Job/Job.csproj
@@ -4,8 +4,4 @@
     <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\Job\Job.csproj" />
-  </ItemGroup>
-
 </Project>
